Add external registrar support and header override
The current implementation did not predict the possibility of an external registrar component (separate from proxy). This commit adds the modification to the SipPhone and SipStack in order to create phones aimed at registering on this external registrar.
Also added the possibility of header overrides within the REGISTER request in order to stress-test the registrar processing.