Substitute loopback with request matching strategy
The current loopback implementation was limited in the sense that it wasn't open to mutation of request matching, causing a SipPhone client to potentially miss an incoming request which did not fit either the current Request URI check or the To check by equality.

This commit introduces a concept of a request matching strategy processing, which is evaluated against the incoming request. This (collection of) strateg(y/ies) provides a determination mechanism of acceptance of request into the SipSession. Thus a user may provide extensions to the SipSession on which inbound requests may be accepted without the need to mutate any parameters. To enable this behavior, some parameters accessors have been moved from SipPhone to its parent, the SipSession.

The request matching mechanism deprecates the loopback accessors. However, the loopback behavior has been retrofitted with the new matching strategy mechanism (tests included for proof).