Refactor request matching logic into a separate object
In order to follow extensibility guidelines defined in the Gang Of Four book on Design Patterns, we refactor the request matching mechanism to increase the cohesion of the SipSession object (which just queries the matching and handles that the matching is up to its business logic). The business logic of the strategy handling themselves is handled now by the request matcher, which will determine if a strategy will be added, removed or contained.

To reduce coupling of the matching strategy to the SipSession, we move the sip session to be an argument of the strategy determination. This way we can represent and initialize each handler as a functional interface (for Java 8), reducing code footprint while still delegating the responsibility of request attribution to a session to the strategy itself.