Add discrimination for singleton-type strategies
The To and Request URI matching strategies are now treated as singleton strategies, permitting only one instance in the matcher. This is to improve performance and remove bugs when adding multiple same-typed strategies. If a strategy wants to be instantiated multiple times (which is default), it has to return the corresponding boolean parameter. This may be applicable in cases where a SipUnit client wants to, for example, accept multiple numbers in the To header using a same strategy class.