Substitute NOTIFY request handling for strategy implementation
In order to support extensible request handling, we introduce the RequestProcessor class, which is tasked with executing a sequence of strategies over an inbound request for a SipListener instance. These strategies execute sequentially until one of the strategies accepts the request. If no strategies were executed or the validation fails, the governing SipPhone will return an appropriate error (which indicates that either the event was not recognized or the subscription was orphaned).